Enlarge Cinder-Volume container
Cinder volume needs temporary space to convert images. This patch
enlarges the default size of the cinder volume container to 105GB. It
also exposes the size in the user configuration. This way a user can
increase the size if they anticipate converting larger images, or reduce
it if they don't require image conversion.

Related: #166